+++ src/main.cpp
@@ -65,6 +65,7 @@ int main(int argc, char** argv) {
             std::string m = next();
             opts.antiAliasing = m == "off" ? 0 : m == "fxaa" ? 1 : 2;
         }
+        else if (a == "--view") opts.cameraView = next();
         else if (a == "--bind-pose") opts.bindPose = true;
         else if (a == "--stance") opts.stance = next() == "goofy" ? 1 : 0;
         else if (a == "--help" || a == "-h") {
@@ -72,7 +73,7 @@ int main(int argc, char** argv) {
                     "            [--menu main|play|map|rider|scooter[:category]|settings[:tab]|challenges]\n"
                     "            [--windowed|--fullscreen] [--size WxH] [--novsync] [--gpu direct3d12|vulkan] [--gpu-debug]\n"
                     "            [--lang en|ja] [--stance regular|goofy] [--pad xbox|ps|switch] [--no-rider]\n"
-                    "            [--aa off|fxaa|taa]\n"
+                    "            [--aa off|fxaa|taa] [--view third|close|far|first]\n"
                     "            [--env preset] [--autotest test.json] [--screenshot file.png frame] [--camera x,y,z,tx,ty,tz]\n"
                     "            [--debugview n]");
             return 0;